your-job-offer-web/backend/requirements.txt
Flask == 3.0.3
Flask-Cors == 5.0.0
Flask-JWT-Extended == 4.7.1
python-dotenv == 1.0.1